requirements.txt
# app
fastapi==0.111.0
gunicorn==20.1.0
structlog==24.2.0
uvicorn[standard]==0.30.1

# telemetry
opentelemetry-sdk==1.25.0
opentelemetry-api==1.25.0
opentelemetry-exporter-gcp-trace==1.6.0
opentelemetry-resourcedetector-gcp==1.6.0a0
opentelemetry-instrumentation-fastapi==0.46b0